+++ common/src/main/java/de/maxhenkel/voicechat/gui/VoiceChatScreen.java
@@ -9,6 +9,8 @@ import de.maxhenkel.voicechat.gui.tooltips.DisableTooltipSupplier;
 import de.maxhenkel.voicechat.gui.tooltips.HideTooltipSupplier;
 import de.maxhenkel.voicechat.gui.tooltips.MuteTooltipSupplier;
 import de.maxhenkel.voicechat.gui.tooltips.RecordingTooltipSupplier;
+import de.maxhenkel.voicechat.gui.volume.AdjustVolumesScreen;
+import de.maxhenkel.voicechat.gui.widgets.ImageButton;
 import de.maxhenkel.voicechat.gui.widgets.ToggleImageButton;
 import de.maxhenkel.voicechat.intercompatibility.ClientCompatibilityManager;
 import de.maxhenkel.voicechat.voice.client.*;
@@ -26,11 +28,13 @@ public class VoiceChatScreen extends VoiceChatScreenBase {
     private static final ResourceLocation TEXTURE = new ResourceLocation(Voicechat.MODID, "textures/gui/gui_voicechat.png");
     private static final ResourceLocation MICROPHONE = new ResourceLocation(Voicechat.MODID, "textures/icons/microphone_button.png");
     private static final ResourceLocation HIDE = new ResourceLocation(Voicechat.MODID, "textures/icons/hide_button.png");
+    private static final ResourceLocation VOLUMES = new ResourceLocation(Voicechat.MODID, "textures/icons/adjust_volumes.png");
     private static final ResourceLocation SPEAKER = new ResourceLocation(Voicechat.MODID, "textures/icons/speaker_button.png");
     private static final ResourceLocation RECORD = new ResourceLocation(Voicechat.MODID, "textures/icons/record_button.png");
     private static final ITextComponent TITLE = new TranslationTextComponent("gui.voicechat.voice_chat.title");
     private static final ITextComponent SETTINGS = new TranslationTextComponent("message.voicechat.settings");
     private static final ITextComponent GROUP = new TranslationTextComponent("message.voicechat.group");
+    private static final ITextComponent ADJUST_PLAYER_VOLUMES = new TranslationTextComponent("message.voicechat.adjust_volumes");
 
     private ToggleImageButton mute;
     private ToggleImageButton disable;
@@ -57,6 +61,13 @@ public class VoiceChatScreen extends VoiceChatScreenBase {
             stateManager.setDisabled(!stateManager.isDisabled());
         }, new DisableTooltipSupplier(this, stateManager));
         addButton(disable);
+
+        ImageButton volumes = new ImageButton(guiLeft + 6 + 20 + 2 + 20 + 2, guiTop + ySize - 6 - 20, VOLUMES, button -> {
+            minecraft.setScreen(new AdjustVolumesScreen());
+        }, (button, matrices, mouseX, mouseY) -> {
+            renderTooltip(matrices, ADJUST_PLAYER_VOLUMES, mouseX, mouseY);
+        });
+        addRenderableWidget(volumes);
 
         if (client != null) {
             if (client.getRecorder() != null || (client.getConnection() != null && client.getConnection().getData().allowRecording())) {
@@ -85,7 +96,7 @@ public class VoiceChatScreen extends VoiceChatScreenBase {
         addButton(group);
 
         group.active = client != null && client.getConnection() != null && client.getConnection().getData().groupsEnabled();
-        recordingHoverArea = new HoverArea(6 + 20 + 2 + 20 + 2, ySize - 6 - 20, xSize - (6 + 20 + 2 + 20 + 2) * 2, 20);
+        recordingHoverArea = new HoverArea(6 + 20 + 2 + 20 + 2 + 20 + 2, ySize - 6 - 20, xSize - ((6 + 20 + 2 + 20 + 2) * 2 + 20 + 2), 20);
 
         checkButtons();
     }
@@ -133,7 +144,7 @@ public class VoiceChatScreen extends VoiceChatScreenBase {
         if (client != null && client.getRecorder() != null) {
             AudioRecorder recorder = client.getRecorder();
             StringTextComponent time = new StringTextComponent(recorder.getDuration());
-            font.draw(poseStack, time.withStyle(TextFormatting.DARK_RED), (float) (guiLeft + (xSize - font.width(time)) / 2), guiTop + ySize - font.lineHeight - 7, 0);
+            font.draw(poseStack, time.withStyle(TextFormatting.DARK_RED), guiLeft + recordingHoverArea.getPosX() + recordingHoverArea.getWidth() / 2F - font.width(time) / 2F, guiTop + recordingHoverArea.getPosY() + recordingHoverArea.getHeight() / 2F - font.lineHeight / 2F, 0);
 
             if (recordingHoverArea.isHovered(guiLeft, guiTop, mouseX, mouseY)) {
                 renderTooltip(poseStack, new TranslationTextComponent("message.voicechat.storage_size", recorder.getStorage()), mouseX, mouseY);